Update OsuContextMenu in line with framework.
